Fix D2: UB-free heartbeat time decode for all 32-bit values
readTimeFromHeartbeat shifted raw bytes that promote to signed int;
'data[0] << 24' with a top-bit-set byte lands in the sign bit, which is
undefined behavior in pre-C++14 modes. The SAMD node target builds under
gnu++11 and was genuinely affected; the C++14 host build and gnu++17 STM32
targets already defined this, which is why UBSan never flagged it. Cast
each byte to uint32_t before shifting - bit-identical results, but
unconditionally well-defined in every language mode.

Adds boundary round-trip tests at 0, 1, 0x7FFFFFFF, 0x80000000, and
0xFFFFFFFF. Note (deviation from the task's prediction): these pass even
pre-fix on the host because C++14 defines the shift - the fix's value is
standards-conformance on the gnu++11 target, not a host-observable bug.

Spec: specs/002-fix-audit-findings (FR-003). Reviewed by verifier agent:
one comment-wording nit (fixed - the UB claim is scoped to the SAMD
toolchain; STM32 targets build gnu++17); it verified each toolchain's
actual -std flag from the compiler invocations.

+++ lib/radio/Radio.cpp
@@ -42,10 +42,13 @@ uint32_t RadioPacket::readTimeFromHeartbeat() const {
   assert(this->type == HEARTBEAT);
   assert(this->dataLength == 4);
 #endif
-  uint32_t time = this->data[3];
-  time |= this->data[2] << 8;
-  time |= this->data[1] << 16;
-  time |= this->data[0] << 24;
+  // Cast before shifting: uint8_t promotes to (signed) int, and shifting a
+  // top-bit-set byte into the sign bit is undefined behavior in pre-C++14
+  // language modes (e.g. the SAMD node target's gnu++11); C++14 defines it.
+  uint32_t time = (uint32_t)this->data[3];
+  time |= (uint32_t)this->data[2] << 8;
+  time |= (uint32_t)this->data[1] << 16;
+  time |= (uint32_t)this->data[0] << 24;
   return time;
 }
